spz_lab1_8/src/Number.java
/*
 * The MIT License
 *
 * Copyright 2017 vladkampov.
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 * THE SOFTWARE.
 */

import java.util.List;
import java.util.ArrayList;

/**
 *
 * @author vladkampov
 */
public class Number {
    private int value;
    private List<Integer> steps = new ArrayList();
    
    Number(int value) {
        this.value = value;
    }

    Number() {
        this.value = 2;
    }

    public int getPoweredValue(int power, int max_steps) {
        int tmp = this.value;
        
        for (int i = 0; i < max_steps; i++) {
            tmp *= tmp;
            this.steps.add(tmp);
        }
        
        return tmp;
    }
    
    public List<Integer> getSteps() {
        return this.steps;
    }
    
    public int getValue() {
        return value;
    }

    public void setValue(int value) {
        this.value = value;
    }
}
